Change currency types, delete Currency class, add monetary dependency

+++ src/main/java/com/parkingapp/parkingservice/domain/common/Location.java
@@ -3,9 +3,11 @@ package com.parkingapp.parkingservice.domain.common;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class Location {
-    private double latitude;
-    private double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 }
